+++ library/Morloc/Namespace.hs
@@ -401,7 +401,7 @@ data TypeError
   | EmptyExpression EVar
   | MissingFeature Text
   | InfiniteRecursion
-  | FunctionSerialization
+  | FunctionSerialization EVar
 
 data MorlocError
   -- | An error that is associated with an expression index