+++ test-suite/UnitTypeTests.hs
@@ -6,7 +6,6 @@ module UnitTypeTests
   , unitTypeTests
   , typeOrderTests
   , typeAliasTests
-  , jsontype2jsonTests
   , packerTests
   , recordAccessTests
   , whereTests
@@ -18,7 +17,6 @@ module UnitTypeTests
 import Morloc.Frontend.Namespace
 import Morloc.CodeGenerator.Namespace
 import Text.RawString.QQ
-import Morloc.CodeGenerator.Grammars.Common (jsontype2json)
 import qualified Morloc.Data.Doc as Doc
 import Morloc (typecheck, typecheckFrontend)
 import qualified Morloc.Monad as MM
@@ -392,37 +390,6 @@ packerTests =
 --             ]
 --         )
 --     ]
-
-jsontype2jsonTests =
-  testGroup
-    "Test conversion of JsonType's to JSON text"
-    [ jsontest "value"
-        (VarJ "int")
-        [r|"int"|]
-    , jsontest "array(value)"
-        (ArrJ "list" [VarJ "int"])
-        [r|{"list":["int"]}|]
-    , jsontest "object(value)"
-        (NamJ "Person" [("name", VarJ "Str"), ("age", VarJ "Int")])
-        [r|{"Person":{"name":"Str","age":"Int"}}|]
-    , jsontest "array(array)"
-        (ArrJ "list" [ArrJ "matrix" [VarJ "int"]])
-        [r|{"list":[{"matrix":["int"]}]}|]
-    , jsontest "array(object)"
-        (ArrJ "list" [(NamJ "Person" [("name", VarJ "Str"), ("age", VarJ "Int")])])
-        [r|{"list":[{"Person":{"name":"Str","age":"Int"}}]}|]
-    , jsontest "object(array)"
-        (NamJ "Person" [("name", VarJ "Str"), ("friends", ArrJ "list" [VarJ "Str"])])
-        [r|{"Person":{"name":"Str","friends":{"list":["Str"]}}}|]
-    , jsontest "object(object)"
-        (NamJ "Person"
-          [ ("name", VarJ "Str")
-          , ("pet", NamJ "Animal" [("name", VarJ "Str"), ("species", VarJ "Str")])
-          ])
-        [r|{"Person":{"name":"Str","pet":{"Animal":{"name":"Str","species":"Str"}}}}|]
-    ]
-  where
-    jsontest msg t j = testEqual msg (Doc.render $ jsontype2json t) j
 
 typeAliasTests =
   testGroup