+++ test-suite/golden-tests/interop-7/foo.R
@@ -1,0 +1,3 @@
+f <- function(x){
+  list("asdf", 4.2, x)
+}