+++ ChangeLog.md
@@ -69,6 +69,7 @@ handling for several very different languages (proofs-of-concept).
  * Partially fix naming conflict in Python and R pools caused by use of "f". My
    solution was name mangling, though we need a more permanent solution to our
    problem.
+ * Let user write correct "list" R types for lists, tuples, and records
  [ ] Fix infinite loop raised from `type Cpp Unit = "Unit"` - see case study bug in nj code.
      This problem goes away when I add the namespace (i.e., "mlc:Unit"), but
      still infinite loops are bad.